fix button recover password and fix class
