+++ app/src/main/java/com/homies/homies/services/Adaptador.java
@@ -56,5 +56,4 @@ public class Adaptador extends FragmentPagerAdapter {
         }
         return titulo;
     }
-
 }
