+++ app/src/main/res/drawable/btn_custom.xml
@@ -3,7 +3,7 @@
     android:padding="10dp"
     android:shape="rectangle" >
 
-    <solid android:color="@color/secondaryColor" />
+    <solid android:color="@color/primaryColor" />
 
     <corners
         android:bottomLeftRadius="5dp"
