+++ app/src/main/res/layout/activity_layout_botton_sheet.xml
@@ -49,7 +49,7 @@
             app:layout_constraintTop_toBottomOf="@+id/ingressMail" />
 
         <Button
-            android:id="@+id/recoverButton"
+            android:id="@+id/btnRecover"
             android:layout_width="300dp"
             android:layout_height="wrap_content"
             android:layout_centerInParent="true"
@@ -64,7 +64,7 @@
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@+id/recoverInput" />
+            app:layout_constraintTop_toBottomOf="@+id/recoverInput"/>
 
     </androidx.constraintlayout.widget.ConstraintLayout>
 </ScrollView>
